Created the sprites for the numbers 0-9 and made some memory usage optimization. Mostly is was using smaller data types for integers

+++ Arduino/RocketLaunch/SpriteViewer.h
@@ -7,7 +7,10 @@ struct Sprite
 {
   MatrixUtil::XY spriteSize;
   CHSV colorHSV;
-  uint32_t* spriteMask; // Bitmask per row
+
+  // Bitmask per row
+  // A sprite may only have a maximum of 8 columns to reduce memory usage
+  uint8_t* spriteMask; 
 };
 
 class SpriteViewer
